Integrated new properties to replace constants

+++ src/main/kotlin/io/craigmiller160/authserver/config/TokenAuthorizationConfig.kt
@@ -1,6 +1,6 @@
 package io.craigmiller160.authserver.config
 
 class TokenAuthorizationConfig {
-  var accessTokenExp: Long = 0
-  var refreshTokenExp: Long = 0
+  var accessTokenExp: Int = 0
+  var refreshTokenExp: Int = 0
 }
